webapp/requirements.txt
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.5.3
chardet==2.2.1
CherryPy==8.7.0
configobj==4.7.2
coverage==3.6b3
cx-Oracle==5.2.1
decorator==4.0.10
defusedxml==0.4.1
di==0.3
Django==1.10.5
django-admin-tools==0.8.0
django-auth-ldap==1.2.8
django-bootstrap-breadcrumbs==0.8.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==7.1.0
django-calendarium==1.3.4
django-ckeditor==5.1.1
django-crispy-forms==1.6.1
django-dashing==0.3.2
django-debug-toolbar==1.6
django-extensions==1.7.5
django-filer==1.2.5
django-filter==1.0.1
django-fsm==2.4.0
django-generic-scaffold==0.4.0
django-genericadmin==0.6.1
django-haystack==2.6.0
django-libs==1.67.9
django-menu==0.1.9
django-menus==1.1.2
django-mptt==0.8.7
django-mptt-admin==0.4.2
django-mptt-tree-editor==0.1.2
django-nvd3==0.9.7
django-pdb==0.5.1
django-polymorphic==1.0.2
django-polymorphic-tree==1.3.1
django-realtime==1.1
django-sitetree==1.7.0
django-smuggler==0.8.0
django-tables2==1.2.9
django-tag-parser==3.1
django-taggit==0.21.3
django-treemenus==0.9.2
django-viewflow==0.10.1
django-webtest==1.8.0
django-xworkflows==0.10.1
easy-thumbnails==2.3
enum34==1.1.6
feedparser==5.2.1
funcsigs==1.0.2
future==0.16.0
iniparse==0.4
IPy==0.75
ipython==5.1.0
ipython-genutils==0.1.0
Jinja2==2.8.1
kitchen==1.1.1
MarkupSafe==0.23
mock==2.0.0
olefile==0.43
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
python-ldap==2.4.28
python-nvd3==0.14.2
python-slugify==1.2.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
treebeard==0.1.1
Unidecode==0.4.19
waitress==1.0.1
wcwidth==0.1.7
WebOb==1.7.0
WebTest==2.0.24
XlsxWriter==0.9.6
xworkflows==1.0.4
